+++ Tests/Compiler/AutowirePassTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\DependencyInjection\Tests\Compiler;
 
 use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\IgnoreDeprecations;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
@@ -36,6 +37,7 @@ use Symfony\Component\DependencyInjection\Tests\Fixtures\BarInterface;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\CaseSensitiveClass;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\includes\FooVariadic;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\OptionalParameter;
+use Symfony\Component\DependencyInjection\Tests\Fixtures\WithoutTarget;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\WithTarget;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\WithTargetAnonymous;
 use Symfony\Component\DependencyInjection\TypedReference;
@@ -267,7 +269,9 @@ class AutowirePassTest extends TestCase
         $pass->process($container);
     }
 
-    public function testGuessableUnionType()
+    #[Group('legacy')]
+    #[IgnoreDeprecations]
+    public function testLegacyGuessableUnionType()
     {
         $container = new ContainerBuilder();
 
@@ -276,6 +280,23 @@ class AutowirePassTest extends TestCase
         $container->setAlias(CollisionB::class.' $collision', 'b');
 
         $aDefinition = $container->register('a', UnionClasses::class);
+        $aDefinition->setAutowired(true);
+
+        $pass = new AutowirePass();
+        $pass->process($container);
+
+        $this->assertSame('b', (string) $aDefinition->getArgument(0));
+    }
+
+    public function testGuessableUnionType()
+    {
+        $container = new ContainerBuilder();
+
+        $container->register('b', \stdClass::class);
+        $container->setAlias('.'.CollisionA::class.' $collision', 'b');
+        $container->setAlias('.'.CollisionB::class.' $collision', 'b');
+
+        $aDefinition = $container->register('a', UnionClassesWithTarget::class);
         $aDefinition->setAutowired(true);
 
         $pass = new AutowirePass();
@@ -1077,7 +1098,9 @@ class AutowirePassTest extends TestCase
         $this->assertSame(['Cannot autowire service "some_locator": it has type "Symfony\Component\DependencyInjection\Tests\Compiler\MissingClass" but this class was not found.'], $container->getDefinition('.errored.some_locator.'.MissingClass::class)->getErrors());
     }
 
-    public function testNamedArgumentAliasResolveCollisions()
+    #[Group('legacy')]
+    #[IgnoreDeprecations]
+    public function testLegacyNamedArgumentAliasResolveCollisions()
     {
         $container = new ContainerBuilder();
 
@@ -1085,6 +1108,31 @@ class AutowirePassTest extends TestCase
         $container->register('c2', CollisionB::class);
         $container->setAlias(CollisionInterface::class.' $collision', 'c2');
         $aDefinition = $container->register('setter_injection_collision', SetterInjectionCollision::class);
+        $aDefinition->setAutowired(true);
+
+        (new AutowireRequiredMethodsPass())->process($container);
+
+        $pass = new AutowirePass();
+
+        $pass->process($container);
+
+        $expected = [
+            [
+                'setMultipleInstancesForOneArg',
+                [new TypedReference(CollisionInterface::class.' $collision', CollisionInterface::class)],
+            ],
+        ];
+        $this->assertEquals($expected, $container->getDefinition('setter_injection_collision')->getMethodCalls());
+    }
+
+    public function testNamedArgumentAliasResolveCollisions()
+    {
+        $container = new ContainerBuilder();
+
+        $container->register('c1', CollisionA::class);
+        $container->register('c2', CollisionB::class);
+        $container->setAlias(CollisionInterface::class.' $collision', 'c2');
+        $aDefinition = $container->register('setter_injection_collision', SetterInjectionCollisionWithTarget::class);
         $aDefinition->setAutowired(true);
 
         (new AutowireRequiredMethodsPass())->process($container);
@@ -1160,6 +1208,22 @@ class AutowirePassTest extends TestCase
         (new AutowirePass())->process($container);
     }
 
+    #[Group('legacy')]
+    #[IgnoreDeprecations]
+    public function testNamedAliasByParameterNameIsDeprecated()
+    {
+        $container = new ContainerBuilder();
+
+        $container->register('image.storage', BarInterface::class);
+        $container->registerAliasForArgument('image.storage', BarInterface::class, 'imageStorage');
+        $container->register('without_target', WithoutTarget::class)
+            ->setAutowired(true);
+
+        $this->expectUserDeprecationMessage('Since symfony/dependency-injection 8.1: Relying solely on the name of parameter "$imageStorage" of "Symfony\\Component\\DependencyInjection\\Tests\\Fixtures\\WithoutTarget::__construct()" to match a named autowiring alias is deprecated; use the "#[Target]" attribute.');
+
+        (new AutowirePass())->process($container);
+    }
+
     public function testArgumentWithIdTarget()
     {
         $container = new ContainerBuilder();
@@ -1208,7 +1272,9 @@ class AutowirePassTest extends TestCase
         $this->assertEquals([new TypedReference(A::class, A::class), 'abc'], $container->getDefinition('foo')->getArguments());
     }
 
-    public function testAutowireUnderscoreNamedArgument()
+    #[Group('legacy')]
+    #[IgnoreDeprecations]
+    public function testLegacyAutowireUnderscoreNamedArgument()
     {
         $container = new ContainerBuilder();
 
@@ -1218,6 +1284,18 @@ class AutowirePassTest extends TestCase
         (new AutowirePass())->process($container);
 
         $this->assertInstanceOf(\DateTimeImmutable::class, $container->get('foo')->now_datetime);
+    }
+
+    public function testAutowireUnderscoreNamedArgument()
+    {
+        $container = new ContainerBuilder();
+
+        $container->autowire(\DateTimeImmutable::class.' $now_datetime', \DateTimeImmutable::class);
+        $container->autowire('foo', UnderscoreNamedArgumentWithTarget::class)->setPublic(true);
+
+        (new AutowirePass())->process($container);
+
+        $this->assertInstanceOf(\DateTimeImmutable::class, $container->get('foo')->dt);
     }
 
     public function testAutowireDefaultValueParametersLike()